Provide command line argument to inform KUBECONFIG path to use.
Fixes https://github.com/gopaddle-io/configurator/issues/32

Allows the controller to receive KUBECONFIG file path
in the form of `--kubeconfig=/path/to/my/kube.config`.

That, in turn, allows developers to run the controller locally without
deploying it to the remote cluster, providing a better experience
debugging a local session in the IDE of their own choice.

Indirectly, it also provides single point where `rest.Config` is created
and reused throughout the code, eliminating the need to reload the
cluster configuration in various points.